refactoring step2
