+++ app/src/main/res/layout/activity_iconotification.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.vladislav.screen.detailscreen.ICONotification">
+    tools:context="com.example.vladislav.screen.detailscreen.ICONotificationActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
